Move TestLoadPackage into parse_test.go
